+++ CHANGELOG.md
@@ -11,6 +11,9 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Allow multiple data types within the message body.
 - Added multiple routing keys to a single queue for monitoring.
 
+### Added
+- \_process_queue:  Private function to process message queue.
+
 ### Changed
 - monitor_queue:  Refactored function to initialize and monitor multiple queues.
 - callback:  Added routing key information to log entries.